Allow keeping temporary schema via `keep_temp_schema` option
For debugging purposes, users can add this configuration to set
the request to keep temporary schema, like the following:

```
db_upload_complete_sql <<EOT
SELECT bde_control.bde_SetOption({{id}},'keep_temp_schema','yes');
EOT
```

Includes automated test

Closes https://github.com/linz/linz-lds-bde-schema/issues/23

+++ sql/02-bde_control_functions.sql
@@ -399,9 +399,13 @@ BEGIN
     WHERE
         upl_id_lock = p_upload;
 
-    -- Drop the schema used by the lock
+    -- Drop the schema used by the lock, unless asked to keep it
 
-    EXECUTE 'DROP SCHEMA IF EXISTS ' || bde_TmpSchema(p_upload) || ' CASCADE';
+    IF bde_GetOption(p_upload, 'keep_temp_schema')::bool THEN
+        RAISE NOTICE 'Keeping temporary schema % as requested by options', bde_TmpSchema(p_upload);
+    ELSE
+        EXECUTE 'DROP SCHEMA IF EXISTS ' || bde_TmpSchema(p_upload) || ' CASCADE';
+    END IF;
 END
 $body$
 LANGUAGE plpgsql;
